Consolidate fs utils into single file
