--- packages/create-toolpad-app/src/util.ts
@@ -1,6 +1,0 @@
-import * as fs from 'fs/promises';
-
-export async function readJsonFile(filePath: string): Promise<any> {
-  const content = await fs.readFile(filePath, { encoding: 'utf-8' });
-  return JSON.parse(content);
-}
