+++ packages/toolpad-app/src/server/FunctionsManager.ts
@@ -6,9 +6,9 @@ import * as chokidar from 'chokidar';
 import chalk from 'chalk';
 import { glob } from 'glob';
 import * as fs from 'fs/promises';
+import { writeFileRecursive, fileExists } from '@mui/toolpad-utils/fs';
 import EnvManager from './EnvManager';
 import { ProjectEvents, ToolpadProjectOptions } from '../types';
-import { writeFileRecursive, fileExists } from '../utils/fs';
 import { createWorker as createDevWorker } from './functionsDevWorker';
 
 const DEFAULT_FUNCTIONS_FILE_CONTENT = `// Toolpad queries:
