+++ packages/toolpad-app/src/server/versionInfo.ts
@@ -1,9 +1,9 @@
 import latestVersion from 'latest-version';
 import * as semver from 'semver';
 import * as path from 'path';
+import { fileExists } from '@mui/toolpad-utils/fs';
 import pkg from '../../package.json';
 import { VERSION_CHECK_INTERVAL } from '../constants';
-import { fileExists } from '../utils/fs';
 import { getUserProjectRoot } from './localMode';
 
 export type PackageManager = 'npm' | 'yarn' | 'pnpm';
